==> CRISPRclean/notebooks/PBMC/1C_Seurat_ref_map.Rmd <==
---
title: "GEX map reference Seurat"
description: "Map PBMC clusters to reference with Seurat"
author: "Daniel Brown"
date: "`r Sys.Date()`"
output:
  html_document:
    toc: true
    toc_depth: 3
    toc_float:
      smooth_scroll: true
    theme: readable
    highlight: tango 
    df_print: paged
    code_folding: hide
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE,
                      dev = "png",
                      dpi = 300,
                      fig.asp=0.6, 
                      fig.path='./figures/PBMC/srtRefMap/',
                      warning=FALSE, message=FALSE)
knitr::opts_knit$set(root.dir = here::here())
library(Seurat)
library(SeuratDisk)
library(ggplot2)
library(patchwork)
library(knitr)
library(tidyverse)
library(here)
source(here::here("scripts/R/theme_publication.R"))
theme_set(theme_Publication())
```

# Aim

Evaluate jumpcode CRISPRclean single-cell boost kit. This adds more content on top of ribosomal and mitochondrial genes.  
Including Pseudogenes and non variable genes

## This sample

I restrict my analysis to R010_MULTI-Seq_GEX in PBMCs. This sample had the best signal to noise.

I sequenced this particular sample much deeper as part of NN265.  
[ELN entry](https://au-mynotebook.labarchives.com/share/Daniel_RnD_Projects/MzQ3LjF8NjgxOTcvMjY3LTE0NzgvVHJlZU5vZGUvMjMwMTMwNzYyN3w4ODEuMDk5OTk5OTk5OTk5OQ==)

## Comparisons

I mapped both the undepleted and CRISPRclean depleted datasets to GRCh38 reference provided by cellranger and the masked reference provided by jumpcode. Jumpcode says only the masked reference will highlight the differences in depleted.

I follow the vignette of [Seurat reference map](https://satijalab.org/seurat/articles/multimodal_reference_mapping.html)

## Aim

Check the cell annotations more comprehensively and choose the best one for presentation.

# Load Data

Read in data

```{r loadData}
sces <- readRDS(here::here(
  "CRISPRclean/data/Jumpcode/SCEs", "PBMC_CRISPR_colData.sces.rds"
))
names(sces)
```

# Convert to Seurat objects

To run reference mapping I need to convert to Seurat

```{r}
srts <- list()

for (n in names(sces)) {
  current <- sces[[n]]
  srt <- Seurat::CreateSeuratObject(counts = counts(current), project = n)
  srt@meta.data <- as.data.frame(colData(current))
  srt <- SCTransform(srt, verbose = TRUE)
  srts[[n]] <- srt
}

names(srts)
```

## Save basic seurat objects

```{r}
saveRDS(srts, here::here(
  "CRISPRclean/data/Jumpcode/SRTs", "PBMC_CRISPR_scTransform.rds"
))
```

# Prepare reference

## A Multimodal PBMC Reference Dataset

We load the reference (download here) from Satija lab recent paper, and visualize the pre-computed UMAP. This reference is stored as an h5Seurat file, a format that enables on-disk storage of multimodal Seurat objects (more details on h5Seurat and SeuratDisk can be found here).

[Link to Seurat azimuth reference](https://atlas.fredhutch.org/data/nygc/multimodal/pbmc_multimodal.h5seurat)

This loading takes a long time to run and many checks are performed.

```{r load_ref}
reference <- LoadH5Seurat(
  "/stornext/HPCScratch/home/brown.d/Projects/R010_multiplexing/data/Azimuth_reference/pbmc_multimodal.h5seurat", verbose=T)

reference
```

# Mapping

## Find anchors {.tabset}
We then find anchors between reference and query. 

### CRISPRclean

```{r anchor_cris}
anchors <- list()

for (n in names(srts)) {
  anchors[[n]] <- FindTransferAnchors(
  reference = reference,
  query = srts[[n]],
  normalization.method = "SCT",
  reference.reduction = "spca",
  dims = 1:50
  )
}

```

We then transfer cell type labels and protein data from the reference to the query. Additionally, we project the query data onto the UMAP structure of the reference.

```{r map_cris}
for (n in names(srts)) {
  srts[[n]] <- MapQuery(
  anchorset = anchors[[n]],
  query = srts[[n]],
  reference = reference,
  refdata = list(
    celltype.l1 = "celltype.l1",
    celltype.l2 = "celltype.l2",
    predicted_ADT = "ADT"
  ),
  reference.reduction = "spca", 
  reduction.model = "wnn.umap"
  )
}
```

# Save reference mapped seurat objects

```{r}
saveRDS(srts, here::here(
  "CRISPRclean/data/Jumpcode/SRTs", "PBMC_CRISPR_refMapped.rds"
))
```

Add cell type labels to SingleCellExperiment

```{r}
for (n in names(sces)) {
  sce <- sces[[n]]
  srt <- srts[[n]]
  sce$predicted.celltype.l1 <- srt$predicted.celltype.l1
  sce$predicted.celltype.l2 <- srt$predicted.celltype.l2
  sce$predicted.celltype.l1.score <- srt$predicted.celltype.l1.score
  sce$predicted.celltype.l2.score <- srt$predicted.celltype.l2.score
  sces[[n]] <- sce
}
```

Save SCEs

```{r loadData}
saveRDS(sces, here::here(
  "CRISPRclean/data/Jumpcode/SCEs", "PBMC_CRISPR_annotated.sces.rds"
))
```

## Next steps 

Visualize the cell type labels generated with Seurat Azimuth in another notebook.

## Session Info

```{r}
sessionInfo()
```
